Phase Matrix list mode sweeping works
* Tested externally triggered sweeps using a spectrum analyzer.
* Sweeps simultaneously in amplitude and frequency.
* Had to add a "LIST:START 1" command after clearing the list to clear the trigger "buffer"
* Works with RS232 sleep time at 30 ms; faster than this is unreliable.

Jason

+++ devices/PhaseMatrix/src/PhaseMatrixDevice.h
@@ -62,7 +62,7 @@ private:
 	bool measurePower(double& power);
 	void clearList();
 
-	void addListPoint(unsigned point, double dwell, const RawEvent& listEvent);
+	void addListPoint(unsigned long point, double dwell, const RawEvent& listEvent);
 
 	bool checkFrequencyFormat(std::string frequency, std::string& formatErrorMessage = trash);
 	bool checkPowerFormat(std::string power, std::string& formatErrorMessage = trash);
